refactor(core): lift Android/iOS init orchestration into GoogleAdManagerBase
AndroidGoogleAdManager (904 lines pre-review-branch, 626 at this branch's
base) and IosGoogleAdManager (637/636) were two independent implementations
of the same policy: the initialize() attempt-coalescing loop, the applied-
configuration decision, native-init generation tracking, terminal status
publication, revocation purging, and placement registration — differing
only in log-message text and the actual GMA/GAD SDK calls. Every future
fix to that policy had to be made twice, and the files had already needed
a cross-file comment to stay aligned (which had itself gone stale — see
83dd822e on the prior branch).

Extracts GoogleAdManagerBase (commonMain, internal, both managers now
`internal class Android/IosGoogleAdManager : GoogleAdManagerBase()`).
Platform subclasses implement seven hooks: platformTag (the one text
difference in every log line), nativeInitializationDispatcher (Android
Main.immediate vs iOS plain Main — preserved as the unexplained divergence
it is, not silently unified), appId, initializeMobileAdsNative (the actual
SDK call), captureDiagnosticsSnapshotOnMain, configureNativeAdsAfterAccepted
Initialization, configuredNativePolicyOrNull, and onNativeConsentRevoked
(NativeAdManagerImpl.onConsentRevoked isn't on the public NativeAdManager
interface and its concrete type differs per platform generic argument).
registerBanner/registerFullScreenSlot in the base collapse what was 5
near-identical validate+collision-check+getOrPut bodies per platform (10
total) into 2 shared implementations; each platform's factory override is
now just the constructor call, which is the one part that's genuinely
different per platform (Android's slots take an Activity provider and an
audio controller iOS has no equivalent for).

registryLock/presenceLock now use the pre-existing FullScreenStateLock
(commonMain, @InternalAdMobCmpApi) instead of each platform hand-rolling
synchronized(Any())/NSRecursiveLock — found while reading NativeAdManagerImpl,
which already used it for exactly this purpose; this repo didn't need a new
abstraction, just to reuse the one it already had.

One real initialization-order hazard, found and designed around rather than
hit at runtime: the original `init { admissionScope.launch { consent... } }`
block cannot move into the base class as-is, because `consent` is an
abstract member each subclass overrides with an eagerly-constructed
property, and Kotlin runs a base class's own init-time code strictly before
a derived class's property initializers — reading `consent` from the base's
init block would observe it before it exists. Exposed as
`protected fun startAdmissionTracking()` instead, called explicitly from
each subclass's own `init {}` after its own `consent` is constructed.
Similarly, `nativeInitializationScope` is `by lazy` rather than eagerly
built, since it depends on the abstract `nativeInitializationDispatcher`.

Fixed two mistakes caught by compiling immediately rather than assumed
correct: `kotlinx.coroutines.cancellation.CancellationException` doesn't
exist (the real package is `kotlin.coroutines.cancellation`), and
FullScreenStateLock's `@InternalAdMobCmpApi` opt-in was missing from the new
file's `@file:OptIn`. Also: `configureNativeAdsAfterAcceptedInitialization`
needed `internal` visibility, not `protected` — two existing white-box tests
(AndroidGoogleAdManagerNativePolicyTest, IosGoogleAdManagerNativePolicyTest)
call it directly as a same-module test, not from a subclass; and detekt
caught `_status`/`_events` as non-conventional once they needed broader-
than-private visibility (renamed `_events` to `mutableEvents`; `_status`
turned out to be usable nowhere outside the base, so stayed private instead).

Comments in the new code were written in changelog voice in a few spots
("extracted from what were previously two ~900-line files", "preserved
as-is by this extraction", "which used to leave initialize() suspended
forever") — caught in review against this repo's own rule
(admob-cmp/AGENTS.md: comments state invariants, not history; git owns
"what changed and when"). Rewritten to state the current rule instead of
narrating this refactor: e.g. the class KDoc now says "keep the fix at this
altitude" instead of describing where the code came from.

Verified, not asserted:
- Both platforms compile clean (no new warnings beyond pre-existing ones).
- `:admob-cmp-core:allTests` passes on both Android host tests and
  iosSimulatorArm64Test, unchanged from before this refactor — including
  the concurrency-sensitive attempt-coalescing, presence-arbiter, and
  revocation-purge paths.
- detekt clean, zero new baseline entries needed.
- `checkKotlinAbi` passes with NO dump changes — both classes are
  `internal` and were already absent from admob-cmp-core.klib.api, so this
  refactor is provably ABI-neutral, not just assumed so.
- Downstream modules (admob-cmp, admob-cmp-compose, showcase) still compile
  against the refactored core.
- Full `./scripts/release-readiness.sh` — READINESS: PASS, all 9 sections,
  nothing skipped.

Net: 1262 -> 920 lines even before accounting for the base class's fuller
KDoc than either terse original carried.
